refactored renderImage to happen in the appropriate components, and got the pencil tool functionality up and running
