+++ src/components/Canvas/Grid.jsx
@@ -1,16 +1,59 @@
-import React, { useContext } from 'react';
-import { renderImage } from '../../utils/grid';
+import React, { useEffect, useState } from 'react';
+import { createImage } from '../../utils/grid';
 import styles from './Grid.css';
-import { imageContext } from '../../context/ImageContext';
+import Pixl from './Pixl';
+import { v4 as uuid } from 'uuid';
 
-export default function Grid(tool) {
-  const { image, index, setPixelArray } = useContext(imageContext);
+export default function Grid() {
+  const [image, setImage] = useState(createImage('title', 10, 10));
+  const [pixelArray, setPixelArray] = useState([]);
 
-  const pixelArray = renderImage(image);
+  function renderImage(imageObject) {
+    const colorArray = imageObject.colors;
+    const area = imageObject.height * imageObject.width;
+    const image = [];
 
-  const pixelClick = () => {
-    console.log('array', pixelArray[i]);
+    for (let i = 0; i < area; i++) {
+      if (imageObject.height >= imageObject.width) {
+        image.push(
+          <Pixl
+            pixelIndex={i}
+            key={uuid()}
+            background={colorArray[i]}
+            width={`${Math.round(500 / imageObject.height)}px`}
+            height={`${Math.round(500 / imageObject.height)}px`}
+            handleClick={handleClick}
+            index={i}
+          />
+        );
+      } else {
+        image.push(
+          <Pixl
+            pixelIndex={i}
+            key={uuid()}
+            background={colorArray[i]}
+            width={`${Math.round(500 / imageObject.width)}px`}
+            height={`${Math.round(500 / imageObject.width)}px`}
+            handleClick={handleClick}
+            index={i}
+          />
+        );
+      }
+    }
+
+    setPixelArray(image);
+  }
+
+  const handleClick = (index) => {
+    const newImage = { ...image };
+    newImage.colors[index] = 'rgb(0, 0, 0)';
+
+    setImage(newImage);
   };
+
+  useEffect(() => {
+    renderImage(image);
+  }, [image]);
 
   return (
     <div
